重构架构，添加健康检查和生命周期管理功能，优化K8s服务发现
